Important commit!
feature_extraction.py: There was a mistake in saving the ERDS values to the csv (Beta C3 and Alpha C4 were interchanged).
The frequency interval for the alpha band was adjusted (from 8-14 Hz to 8-13 Hz).
ERDS values are now rounded to two decimals.
Else were just small changes not effecting the results.